avoid checking for bootstrap secret and infra ready for control plan
We do not have support for ElasticIP and the modem. Having an IP to set
as APIEndpoint is a requirement for Infra.Ready = true. Without an IP we
can set it.

It means that for now we have to skips a few check in the machine
control loop, but only for control plans, otherwise we will get stuck
waiting for an IP that will never arrive.

This is temporary, as soon as we can we should create an ElasticIP as
part of the cluster controller.

I think this prevents multi control plan to work