Move facility from packetmachine to packetcluster
Facility describes where a cluster lives not a machine because it drives
users to the wrong assumption that they can place machines part of the
same cluster across facilities.

This is not how kubernetes works at the moment. The right solution to do
a multi regional cluster is via federation.

I have also changed type from a slice of string to a string. It will may
work for Packet API but I presume this is not something that applies to
Kubernetes Cluster.

If you want a cluster you want in some particular facility.