Teach ccm what it needs to look after the control plane endpoint
One of the Packet CCM responsability is to check the status of the
Control Plane Endpoint (an elastic ip) keeping it assigned to an healthy
control plane.

In order to do that it needs to know the key=pair used to tag the
elastic ip at creation phase. This PR pass this information to CCM via
Kubernetes Secret