Enable Control Plane elastic ip management in CCM
This PR enables the Elastip IP management for Control Plane in CCM.

One of the Packet CCM responsibility is to health check the Control
Plane IP and move the Elastic IP to an healthy control plane.

In order to do that the configuration has to be set to CCM secret.